# feat: Support read local file async