--- a/velox/common/file/File.cpp
+++ b/velox/common/file/File.cpp
@@ -129,8 +129,11 @@ uint64_t InMemoryWriteFile::size() const {
   return file_->size();
 }
 
-LocalReadFile::LocalReadFile(std::string_view path, bool bufferIo)
-    : path_(path) {
+LocalReadFile::LocalReadFile(
+    std::string_view path,
+    folly::Executor* executor,
+    bool bufferIo)
+    : executor_(executor), path_(path) {
   int32_t flags = O_RDONLY;
 #ifdef linux
   if (!bufferIo) {
@@ -160,7 +163,8 @@ LocalReadFile::LocalReadFile(std::string_view path, bool bufferIo)
   size_ = ret;
 }
 
-LocalReadFile::LocalReadFile(int32_t fd) : fd_(fd) {}
+LocalReadFile::LocalReadFile(int32_t fd, folly::Executor* executor)
+    : executor_(executor), fd_(fd) {}
 
 LocalReadFile::~LocalReadFile() {
   const int ret = close(fd_);
@@ -243,6 +247,20 @@ uint64_t LocalReadFile::preadv(
   }
 
   return totalBytesRead;
+}
+
+folly::SemiFuture<uint64_t> LocalReadFile::preadvAsync(
+    uint64_t offset,
+    const std::vector<folly::Range<char*>>& buffers) const {
+  auto [promise, future] = folly::makePromiseContract<uint64_t>();
+  executor_->add([this,
+                  _promise = std::move(promise),
+                  _offset = offset,
+                  _buffers = buffers]() mutable {
+    auto delegateFuture = ReadFile::preadvAsync(_offset, _buffers);
+    _promise.setTry(std::move(delegateFuture).getTry());
+  });
+  return std::move(future);
 }
 
 uint64_t LocalReadFile::size() const {