@jinchengchenghh 2024-12-16
Linux has aio interface, https://www.man7.org/linux/man-pages/man7/aio.7.html, but is not widely used, not sure which is better. I will also implement mmap to find which one is better. Maybe aio is better when the device is slow.