@Yuhta 2024-12-17
Usually cache is SSD so it's hard to tell which one is faster