@jinchengchenghh 2024-12-18
> Usually cache is SSD so it's hard to tell which one is faster

In that case, we may need mmap, we could add the option to let user decides which to use.

Some one may use HDD. I test the async read in my local SSD environment, it shows time is similar with before.