@jinchengchenghh 2024-12-23
> Let's add options in file system options to enable async read? And only local file system needs to respect that option by creating a executor to support that? Also we shall use cpu executor here instead of io executor

If we add the option, we will also need to set it by QueryConfig in FileInputStream. Do we add this config?  All the file system should respect this config unless the config indicates to it only applies to local file system. Spark only spills to local file, does not support to spill to remote storage. But we can support other file system by creating an executor too. no use case for it now.