@xiaoxmeng 2024-12-24
> Usually cache is SSD so it's hard to tell which one is faster

yeah, for ssd, no need for async reada and it is only supposed to be used on hdd.