@jinchengchenghh 2024-12-30
> @jinchengchenghh I also think this should be an option.
> 
> > If we add the option, we will also need to set it by QueryConfig in FileInputStream.
> 
> Can you explain this further? I thought #11783 changed to use Velox FS. Where should the QueryConfig be set?
> 
> > All the file system should respect this config unless the config indicates to it only applies to local file system.
> 
> Let's name the config to be local file-system specific. All other file-system config names are specific anyway.

Thanks for your reply, discussed with @xiaoxmeng offline, I have changed it to the current version. Add FileSystemOptions to function registerLocalFileSystem, how about that?