@jinchengchenghh 2024-12-31
Current interface `ReadFile` has three public APIs, `preadv`, `preadvAsync` and `hasPreadvAsync`
https://github.com/facebookincubator/velox/blob/main/velox/common/file/File.h#L102,
LocalReadFile just implements it. Would you prefer to move the latter two APIs to private? This may deviate with current design.  CC @xiaoxmeng