@majetideepak 2024-12-31
> Current interface ReadFile has three public APIs, preadv, preadvAsync and hasPreadvAsync

@jinchengchenghh thanks for the clarification. I missed this. Then why do we need an option? Can we implement `preadvAsync` and `hasPreadvAsync` as you did without any options?