@xiaoxmeng 2024-12-31
> > Current interface ReadFile has three public APIs, preadv, preadvAsync and hasPreadvAsync
> 
> @jinchengchenghh thanks for the clarification. I missed this. Then why do we need an option? Can we implement `preadvAsync` and `hasPreadvAsync` as you did without any options?

@majetideepak  The async read support on local file system relies on a cpu executor running at the background for async io (note POSIX has async read support using signal mechanism which is not super reliable or ease to use long years back. Not sure how good it is now). This requires creating a cpu executor inside the local fs instance. The local fs is only used with ssd in Meta internal use cases which don't really need async read support performance wise. But it should be quite useful when used with HDD to pipeline io with computation. So it might be better to have an option for local fs to enable it or not by creating a dedicated cpu executor for that. We could extend local fs registration API to take this option on server startup. For Prestissimo, it could be a system config. thanks!