@majetideepak 2024-12-31
> So it might be better to have an option for local fs to enable it or not by creating a dedicated cpu executor for that.

@xiaoxmeng To clarify, the usage would be `preadv` for SSD and `preadvAsync` for HDD.
Is there any downside if we always create a CPU executor for a LocalReadFile instance?