@kecookier 2025-01-16
@ccat3z Can you post your test results with the HDD here?