@ccat3z 2025-01-16
> @ccat3z Can you post your test results with the HDD here?

I tested some internal jobs on HDD. According to the read spill metrics (spillReadWallNanos and spillReadBytes), async file reading can reduce the file reading time by 52%. However, the time to read spills in these jobs is very small, there's no noticeable speedup in the total time taken.


impl | operator | readTime (Seconds) | spillRead (GB)| Time taken pre GB
-- | -- | -- | -- | --
before | hash agg | 174965.1473 | 113825.6283 |  
after | hash agg | 77624.91614 | 69806.41933 | 72.34%
before | hash build | 52.222391 | 13.31936284 |  
after | hash build | 50.677976 | 13.31939661 | 97.04%
before | hash probe | 401.133286 | 1157.013263 |  
after | hash probe | 199.017824 | 1157.072997 | 49.61%
before | sort | 57569.94362 | 75449.8015 |  
after | sort | 16694.44938 | 75554.49983 | 28.96%
before | total | 232988.4466 | 190445.7624 |  
after | total | 94569.06132 | 146531.3116 | 47.25%