@jinchengchenghh 2025-01-17
https://github.com/facebookincubator/velox/issues/12113
https://github.com/facebookincubator/velox/issues/12100
https://github.com/facebookincubator/velox/issues/12071
The red CI is caused by these 3 issues, not related to this PR.